shows route in RouteDetail
